ticktick/requirements.txt
requests
Flask
python-dotenv
pyinstaller

